+++ src/app/core/models/config.model.ts
@@ -1,0 +1,7 @@
+export class Config {
+    constructor(
+        public id: number,
+        public login: string,
+        public email: string) {
+    }
+}