+++ src/app/products/components/product-list-component/product-list-component.component.ts
@@ -23,7 +23,7 @@ export class ProductListComponentComponent implements OnInit {
   }
 
   public onBuyProduct(product: Product) {
-    this.cartService.add(product);
+    this.cartService.addProduct(product);
   }
 
 }